+++ docs/Changelog.md
@@ -1,7 +1,6 @@
 ### Known issue/TODO
 - Using translation from pointer to audiocategory to get clips
 - Added errortext in PluginConfigurator using ConfigHeader and attribute Archor:UpperLeft
-- Remove CScore for MS-PL
 
 ### v1.0.2
 - Fixed when you quit the mission in the water, the muffle effect still exists when playing audio.
@@ -9,6 +8,7 @@
 - Abstract the audioLoader as a individual class, makes the Grey Announcer become as a framework.
 - Some changes of license (fixed/added the Lib\ folder that does not include in License, This folder includes assembly-sharp of ULTRAKILL, PluginConfigurator and some unity assets).
 - Added parallel audio playing option.
+- Added Load-and-Play loading options for RAM performance.
 - switch to using BepInex's config to customize audio loading path (Default path is the legacy method, you can edit it easily via PluginConfigurator).
 - RECONSTRUCT the code architecture.
 - removed patch to update() in StyleHUD.cs